pelicanconf.py: Change date format to 'yyyy年mm月dd日'
The old date format 'dd日 mm月 yyyy年' gets the characters right for
western dates in Chinese and Japanese, but they would never be done
in that order; in China and Japan dates are invariably given with the
year first, then month, then day. (And without spaces.)

This also ensures that even for those who cannot read the characters
the components will be obvious, since they need not figure out whether
the year-last format is European (DD-MM-YYYY) or American (MM-DD-YYYY).